pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "evercore"
version = "0.1.0"
description = "Standalone generic ticket/task/worker engine with workflow execution"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.24",
    "sqlalchemy>=2.0.0",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0.2",
    "pytz>=2025.2",
    "lemlem>=0.1.0",
]

[project.scripts]
evercore-api = "evercore.api:main"
evercore-worker = "evercore.worker:main"

[tool.hatch.build.targets.wheel]
packages = ["src/evercore"]

[tool.uv.sources]
lemlem = { workspace = true }

[tool.ruff]
ignore = ["E501"]
